Make storage more compact
This commit is a relatively big overhaul of the small code base.
Basically fbitset is essentially an arbitrary-precision dynamic bit set
with short-string optimization.   However, in the previous code, for
simplicity, the in-place storage and the out-of-place storage are laid
out side-by-side, even only one of them could be in use.

In this commit, a raw union type is used to store the contents more
compactly.  This introduces significant burden for the resource
management.  As a result, the core storage and resource management is
refactored into a separate base class.