Remove duplicated number of limbs utility
The utility for computing the number of limbs required for a bit set
with a given size was actually duplicated during the separation of the
core storage and the extra functionalities.  This commit fixes this
issue.  Since this is useful even for the base, the one in the base was
kept.