Move core access methods to the base class
In the initial overhaul, the internal storage union was marked as
`protected` in order for the actual bit set class to work, which does
not look a good design.

Now the core accessors are moved inside the base, so the union thing can
be handled in a completely transparent way by the subclasses.  These
accessors are put as protected, with the data now marked as private.