Fix optimization by loop unrolling
After the big update for more compact storage in commit db053b8, the
`n_limbs` method actually changed its semantics.  Previously, for limbs
stored in-place, this method always returns the size of the in-place
storage.  However, now the minimum number of limbs to hold the size is
returned for both storage mode.  In this way, its value is no longer an
effective constexpr any more.  After inspection of the assembly code
from the compiler, indeed the correct loop unrolling is gone.

Luckily, for std::array types, it is found that range-based for loop or
loop based on the `size()` method seems to work.  As a result, these
paradigms are now used and advocated.