Clarify impact of size on equality and hashing
Now with a slightly different concept of size, this commit makes it
clear that bit sets constructed to be different sizes will be considered
unequal and hash differently, even when they have the same set of bits
set to true.

The code for the hashing has also been optimized and made clearer.