Rework the iterator internals
Previously, the iterators keep a *whole copy* of the entire fbitset
during its construction, and gradually mutate all the bits in it to
zero.  This clearly is not the best way to do this.

Now the iterator keeps a pair of pointers into the limbs, and only copy
and mutate the current limb.  This could avoid possible memory
allocations when external storage of the limbs is needed, and also keeps
all iterator operations to be of constant O(1) cost.