Update the Size type definition
Previously, for some misled and obscure reasons, the native size type
for the fbitset library was defined to be `int`.  This could limit the
size to be 512M bytes only.  And it may also implicitly introduce some
unnecessary casting when the native `size_t` is expected.  As a result,
in this commit, it is updated to the native `size_t`.

Although all test passes, all code using this library may have
compatibility issues.  This is the first, and probably only, backward
incompatible change that has ever happened to fbitset.